unify hyperelastic and tensor-based materials
with `stress, statevars_new = umat.gradient([F, statevars])` and `elasticity = umat.hessian([F, statevars])`
